Move repliction event stream handling out of slave store
